Add files via upload
Breakdown of analysis of 'The Prayer for the Jews in the Ordinary Form (1970 Missal, 2011 English Translation)'.
1.	Why should the Jews ‘advance’ if they have rejected the Messiah? And what here means advancement of the Jews? 
The verb ‘to advance’: make or cause to make progress; to go forward purposely.

2.	The stipulations laid down in the prayer for advancement - ‘in love of his name and in faithfulness to his covenant’.
a.	Continue to worship their God (Hashem) - presumably this condition has been satisfied by any practising Jewish person or group. 
b.	Continue to remain faithful to an obsolete covenant- that is the old covenant which has been superseded. Nonetheless, presumably this condition has been satisfied by any practising Jewish person or group. 
So, this should read ‘may the Jews continue to advance while they continue to do what they are currently doing, i.e., deny Christ and fail to accept His ‘new’ covenant.

3.	The promises to Abraham’s descendants remain only if they accept the ‘New’ Covenant.

4.	Praying for the ‘Fullness of redemption’ for the Jewish people without first praying for their conversion, i.e., making it first clear that they accept that that Jesus Christ I.N.R.I is their Lord and Saviour is ‘Perfidia’ that is faithlessness - on the part of the Roman Church apparatus.

5.	The setting in which this prayer is also important. This prayer is recited during the Holy Sacrifice of the Mass, on Good Friday. On Good Friday during the Holy Sacrifice of the Mass, the faithful who recite this prayer ask the ‘Almighty ever-living God’ to grant the intentions of the prayer ‘Through Christ our Lord’.

The question we must ask is whether Almighty God can deny the intentions of this prayer when it invokes:
1. The sacrifice of Christ in the Eucharist. 
2. His own son ‘Christ our Lord’. 
3. The hearts of the agreeing gathered faithful who answer Amen,  i.e., the expression of strong agreement generally adopted in Christian worship as a concluding word.
Accordingly, if this prayer is answered, the Jews are sure to advance with this prayer.